+++ modules/3d-tiles/src/3d-tiles-archive-loader.ts
@@ -1,6 +1,6 @@
 import {LoaderOptions, LoaderWithParser} from '@loaders.gl/loader-utils';
 import {DataViewFile} from '@loaders.gl/zip';
-import {parse3tz as parse3tzFromProvider} from './tz3/tz3-parser';
+import {parse3DTilesArchive as parse3DTilesArchiveFromProvider} from './3d-tiles-archive/3d-tiles-archive-parser';
 
 // __VERSION__ is injected by babel-plugin-version-inline
 // @ts-ignore TS2304: Cannot find name '__VERSION__'.
@@ -24,7 +24,7 @@ export const Tiles3DArchiveFileLoader: LoaderWithParser<
 > = {
   name: '3tz',
   id: '3tz',
-  module: 'i3s',
+  module: '3d-tiles',
   version: VERSION,
   mimeTypes: ['application/octet-stream', 'application/vnd.maxar.archive.3tz+zip'],
   parse: parse3DTilesArchive,
@@ -42,6 +42,6 @@ async function parse3DTilesArchive(
   data: ArrayBuffer,
   options: Tiles3DArchiveFileLoaderOptions = {}
 ): Promise<ArrayBuffer> {
-  const archive = await parse3tzFromProvider(new DataViewFile(new DataView(data)));
+  const archive = await parse3DTilesArchiveFromProvider(new DataViewFile(new DataView(data)));
   return archive.getFile(options.tz3?.path ?? '');
 }
