+++ modules/3d-tiles/src/index.ts
@@ -2,8 +2,8 @@
 export {Tiles3DLoader} from './tiles-3d-loader';
 export {CesiumIonLoader} from './cesium-ion-loader';
 export {Tile3DSubtreeLoader} from './tile-3d-subtree-loader';
-export type {Tiles3DArchiveFileLoaderOptions} from './tz3-loader';
-export {Tiles3DArchiveFileLoader} from './tz3-loader';
+export type {Tiles3DArchiveFileLoaderOptions} from './3d-tiles-archive-loader';
+export {Tiles3DArchiveFileLoader} from './3d-tiles-archive-loader';
 
 // WRITERS
 export {Tile3DWriter} from './tile-3d-writer';
